+++ sdk/identity/azure_identity/src/azure_cli_credential.rs
@@ -47,10 +47,6 @@ impl CliTokenResponse {
 }
 
 impl OutputProcessor for CliTokenResponse {
-    fn credential_name() -> &'static str {
-        "AzureCliCredential"
-    }
-
     fn get_error_message(_stderr: &str) -> Option<String> {
         // Azure CLI's errors are generally clear and more helpful than anything we'd write here
         None
@@ -225,12 +221,14 @@ mod tests {
 
     #[tokio::test]
     async fn error_includes_stderr() {
-        let stderr = "something went wrong";
-        let err = run_test(1, "stdout", stderr, None, None)
+        let err = run_test(1, "stdout", "something went wrong", None, None)
             .await
             .expect_err("expected error");
         assert!(matches!(err.kind(), ErrorKind::Credential));
-        assert!(err.to_string().contains(stderr));
+        assert_eq!(
+            "AzureCliCredential authentication failed. something went wrong\nTo troubleshoot, visit https://aka.ms/azsdk/rust/identity/troubleshoot#azure-cli",
+            err.to_string()
+        );
     }
 
     #[tokio::test]
@@ -259,7 +257,10 @@ mod tests {
             .await
             .expect_err("error");
         assert!(matches!(err.kind(), ErrorKind::Credential));
-        assert!(err.to_string().contains("az login"));
+        assert_eq!(
+            "AzureCliCredential authentication failed. Please run 'az login' to setup account.\nTo troubleshoot, visit https://aka.ms/azsdk/rust/identity/troubleshoot#azure-cli",
+            err.to_string()
+        );
     }
 
     #[tokio::test]